+++ backend/src/main/java/org/example/backend/jwt/LoginFilter.java
@@ -1,4 +1,4 @@
-package org.example.backend.global.config;
+package org.example.backend.jwt;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.HttpServletRequest;
